Refactored parts of swscale, and filter packages
